educonecta/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
google-generativeai==0.8.5
sendgrid==6.11.0
twilio==9.5.2
python-dotenv==1.1.0
